refactor: update global styles and configuration
- Changed global CSS file path from `src/app/globals.css` to `src/app/styles/global.css`.
- Enabled CSS variables in Tailwind configuration.
- Removed unused CSS imports and styles from `global.css`.
- Added new `next.config.ts` file to specify build directory.
- Updated import paths to use the new alias structure for better organization.
- Set background image for Hero component using a new asset.
- Adjusted Navbar and Avatar components to reflect new utility imports.
