--- web/src/app/global.css
@@ -1,9 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-body {
-  background: #04050A;
-  color: white;
-  color-scheme: dark;
-}
