+++ web/src/components/common/Navbar.tsx
@@ -3,11 +3,11 @@
 import React from 'react';
 import Image from 'next/image';
 import { motion } from 'framer-motion';
-import { Logo } from '@public/shared';
+import { Logo } from '@/public/shared';
 import { type NavbarProps } from '@/lib/types';
 import { COMPANY_INFO } from '@/lib/constants';
 import { fadeInDown, hoverScale } from '@/lib/utils/animations';
-import { cn } from '@/lib/utils';
+import { cn } from '@/src/lib/utils';
 
 const Navbar: React.FC<NavbarProps> = ({ className }) => {
   return (
@@ -17,16 +17,9 @@ const Navbar: React.FC<NavbarProps> = ({ className }) => {
       animate="animate"
       className={cn(
         'flex items-center justify-center gap-x-2 sm:gap-x-3',
-        'bg-[#B2E2F0] rounded-[15px] sm:rounded-[20px] md:rounded-[25px]',
-        'px-3 sm:px-4 md:px-6 py-2 sm:py-3 md:py-4',
-        'shadow-md mt-2 sm:mt-4 md:mt-8 mx-auto',
-        'min-w-[250px] min-h-[60px]',
+        'mt-2 sm:mt-4 md:mt-8 mx-auto',
         className
       )}
-      style={{
-        width: 'min(382.8px, 90vw)',
-        height: 'min(100px, 12vh)',
-      }}
     >
       <motion.div {...hoverScale}>
         <Image
