+++ web/src/components/ui/avatar.tsx
@@ -2,7 +2,7 @@
 
 import * as React from 'react';
 import Image from 'next/image';
-import { cn } from '@/lib/utils';
+import { cn } from '@/src/lib/utils';
 
 interface AvatarProps {
   className?: string;
